Move formatter to SQLX from API
