+++ cli/index.ts
@@ -4,7 +4,7 @@ import * as path from "path";
 import yargs from "yargs";
 
 import * as chokidar from "chokidar";
-import { build, compile, credentials, format, init, install, run, table, test } from "df/api";
+import { build, compile, credentials, init, install, run, table, test } from "df/api";
 import { CREDENTIALS_FILENAME } from "df/api/commands/credentials";
 import * as dbadapters from "df/api/dbadapters";
 import { prettyJsonStringify } from "df/api/utils";
@@ -35,6 +35,7 @@ import { actuallyResolve, assertPathExists, compiledGraphHasErrors } from "df/cl
 import { createYargsCli, INamedOption } from "df/cli/yargswrapper";
 import { supportsCancel, WarehouseType } from "df/core/adapters";
 import { dataform } from "df/protos/ts";
+import { formatFile } from "df/sqlx/format";
 
 const RECOMPILE_DELAY = 500;
 
@@ -591,7 +592,7 @@ export function runCli() {
           const results = await Promise.all(
             filenames.map(async filename => {
               try {
-                await format.formatFile(path.resolve(argv["project-dir"], filename), {
+                await formatFile(path.resolve(argv["project-dir"], filename), {
                   overwriteFile: true
                 });
                 return {
